+++ selftests/core/assets/bkr.json
@@ -1,0 +1,23 @@
+[
+    {
+        "distro_name": "Distro-1.2.0-20230110"
+    },
+    {
+        "distro_name": "Distro-1.2.0-20230109"
+    },
+    {
+        "distro_name": "Distro-1.2.0-20230108"
+    },
+    {
+        "distro_name": "Distro-1.2.0-20230107"
+    },
+    {
+        "distro_name": "Distro-1.2.0-20230106"
+    },
+    {
+        "distro_name": "Distro-1.2.0-20230105"
+    },
+    {
+        "distro_name": "Distro-1.2.0-20230104"
+    }
+]
